tests: Start tests for disks with 4k sector size
We added API for using 4k sector size and various alignments, but there
are not tests for the new APIs yet. Testing 4k block size is more
complicated since creating a block device and mounting file systems
requires root, and we like to run the tests as a regular user.

Add new tests for writing and reading lockspace and resource using 4k
storage provided by the user. If a path is not specified the tests are
skipped.

To run the tests, you need to create a block device using 4k sector size
and optionally create a file system and mount it. The process is
explained in README.dev.

Then pass the path of the block device or file on the new file system to
the tests:

    SANLOCK_4K_PATH=/dev/loop2 tox -e py26

Adding and removing lockspace and acquiring resources is not tested yet
with 4k.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>